requirements.txt
Streamlit 
os
python-dotenv
langchain-google-genai
langchain-community
google-generativeai
PyPDF2
faiss-cpu
